chore(store): move store-related files to store folder and change imports
