+++ src/Lists/ListItem.tsx
@@ -5,7 +5,7 @@ import { useSelector } from 'react-redux';
 /* APPLICATION */
 import edit from '../icons/edit.svg';
 import remove from '../icons/remove.svg';
-import { selectAllCategories } from '../features/categoriesSlice';
+import { selectAllCategories } from '../store/categoriesSlice';
 import { ModalEditItem } from '../Modal/ModalEditItem';
 import { ModalRemoveItem } from '../Modal/ModalRemoveItem';
