+++ src/index.tsx
@@ -5,7 +5,7 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 
 /* APPLICATION */
-import { store } from './app/store';
+import { store } from './store/store';
 import App from './App';
 //import reportWebVitals from './reportWebVitals';
 import './index.css';
