+++ src/store/categoriesSlice.ts
@@ -3,7 +3,7 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { v4 as uuidv4 } from 'uuid';
 
 /* APPLICATION */
-import { RootState } from '../app/store';
+import { RootState } from './store';
 
 export interface CategoriesState {
 	id: string;
@@ -49,7 +49,7 @@ export const categoriesSlice = createSlice({
 			}
 		},
 		categoriesRemoved: (state: CategoriesState[], action: PayloadAction<string>) => {
-			let rm = (el: CategoriesState, i: number, arr: CategoriesState[]) =>
+			const rm = (el: CategoriesState, i: number, arr: CategoriesState[]) =>
 					el.id === action.payload,
 				rmTaskIndex = state.findIndex(rm);
