+++ src/store/store.ts
@@ -1,6 +1,6 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import categoriesReducer from '../features/categoriesSlice';
-import tasksReducer from '../features/tasksSlice';
+import categoriesReducer from './categoriesSlice';
+import tasksReducer from './tasksSlice';
 
 export const store = configureStore({
 	reducer: {
